snapshot funcs
